refactor: Rank vo 로 원상복귀
