+++ src/test/java/lotto/domain/rank/vo/RankTest.java
@@ -1,6 +1,5 @@
-package lotto.domain.rank;
+package lotto.domain.rank.vo;
 
-import lotto.domain.rank.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
